* Fixes for GCC 4.2.4. * Update v097 Scons build environment. * Add missing string API for histogram module. * Fix typo cleaning up clock time update logic. * Remove unused inlining on transmit window also incurring c89 legacy styling.

+++ openpgm/pgm/include/pgm/list.h
@@ -44,7 +44,7 @@ pgm_list_t* pgm_list_prepend_link (pgm_list_t* restrict, pgm_list_t* restrict) P
 pgm_list_t* pgm_list_remove_link (pgm_list_t*, pgm_list_t*) PGM_GNUC_WARN_UNUSED_RESULT;
 pgm_list_t* pgm_list_delete_link (pgm_list_t*, pgm_list_t*) PGM_GNUC_WARN_UNUSED_RESULT;
 pgm_list_t* pgm_list_last (pgm_list_t*);
-size_t pgm_list_length (pgm_list_t*);
+unsigned pgm_list_length (pgm_list_t*);
 
 
 PGM_END_DECLS
